Fix correctness bugs found reviewing c_src
- nif_utils.hpp: error() copied strlen(msg)+1 bytes into a binary that
  enif_make_new_binary sized for strlen(msg); switch to memcpy.
- evision.cpp: evision_from(Point3f) emitted x twice and dropped z.
- evision.cpp: convert_to_erlang_tuple dropped the env argument, so
  evision_from(std::tuple<...>) would fail to compile if instantiated.
- evision.cpp: evision_from for std::vector<char>/String/std::string
  used strncpy, which truncates binary data at embedded NUL bytes;
  switch to memcpy.
- erlcompat.hpp: CV_ERL_TO_ENUM inverted its success check, returning
  false for a valid integer and true (with an uninitialised value)
  otherwise.
- from_binary.h: declared_size overflowed in 32-bit int arithmetic
  before widening to size_t; cast before multiplying.
- windows_fix.cpp: error() built the tuple with the atom() function
  instead of the error_atom value.

+++ c_src/modules/evision_backend/from_binary.h
@@ -36,7 +36,7 @@ static ERL_NIF_TERM evision_cv_mat_from_binary(ErlNifEnv *env, int argc, const E
                 // validate binary data
                 int type = 0;
                 if (!get_binary_type(t, l, img_channels, type)) return evision::nif::error(env, "not implemented for the given type");
-                size_t declared_size = img_rows * img_cols * img_channels * (l/8);
+                size_t declared_size = (size_t)img_rows * img_cols * img_channels * (l / 8);
                 if (declared_size != data.size) return evision::nif::error(env, "size mismatch");
 
                 evision_res<cv::Mat *> * res;
